+++ src/components/Footer/Data.js
@@ -1,0 +1,19 @@
+const footers = [
+    {
+      title: 'Online shop Company',
+      description: [ {desc : 'About us' , link: '/about'}, {desc:'Privacy and policy', link: '/privacy'},{ desc: 'Terms of use', link: '/terms'}]
+    },
+    {
+      title: 'Features',
+      description: [ {desc: 'Best price' , link:'/shop'},{desc: 'Easy shoping',link:'/shop'},{ desc: 'Best products', link:'/shop'}]
+    },
+    {
+      title: 'Social',
+      description: [{desc:'Instagram',link:'/'},{desc: 'Facebook', link:'/'},{desc: 'Twitter',link:'/'}]
+    },
+    {
+      title: 'Get in touch',
+      description: [ {desc:'+123 456 789' , link:'/'},{desc: 'support@onlineshop.ba',link:'/'} ]
+    }
+  ];
+  export default footers;
