+++ src/components/Footer/style.scss
@@ -1,0 +1,13 @@
+* {
+    ul  {
+      margin: 0px;
+      padding: 0px;
+      list-style: none
+    }
+    li {
+      cursor:pointer
+    }
+    a {
+      text-decoration: none
+    }
+  }
